requirements.txt
torch
torchvision
ultralytics
av
numpy
Pillow
